# Performance improvements on the dtrace collapser