--- a/src/collapse/perf.rs
+++ b/src/collapse/perf.rs
@@ -394,10 +394,10 @@ impl Folder {
     }
 
     fn finish<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        let mut keys: Vec<_> = self.occurrences.keys().collect();
+        let mut keys: Vec<_> = self.occurrences.iter().collect();
         keys.sort();
-        for key in keys {
-            writeln!(writer, "{} {}", key, self.occurrences[key])?;
+        for (key, value) in keys {
+            writeln!(writer, "{} {}", key, value)?;
         }
         Ok(())
     }