Each change is one commit with perf delta appended.
Overall towards master:

```
collapse/dtrace         time:   [11.244 ms 11.398 ms 11.585 ms]
                        thrpt:  [113.80 MiB/s 115.68 MiB/s 117.25 MiB/s]
                 change:
                        time:   [-39.575% -37.504% -35.438%] (p = 0.00 < 0.05)
                        thrpt:  [+54.891% +60.011% +65.493%]
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
```