@jonhoo 2019-03-15
I kind of want to see some more of this (`has_inline` in particular) applied to the `perf` collapser too. We could probably tease out a decent bit more performance there as well :D